feat: udpated bathroom filter and make a new page for commercial propertys
